[#221] update unit tests as the get_html_email_options interface changed
